Back/requirements.txt
fastapi
uvicorn[standard]
openai
SpeechRecognition
SQLAlchemy
psycopg2-binary
python-dotenv
rapidfuzz
sounddevice
scipy
pydub
fuzzywuzzy
python-Levenshtein
python-multipart
openai-whisper
soundfile
numpy
fastapi[all]
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.1.3
cryptography
pyzipper